+++ styles/globals.css
@@ -18,7 +18,7 @@
   content: "";
   rotate: 180deg;
   position: absolute;
-  top: -4000px;
+  top: -3000px;
   left: 0;
   right: 0;
   bottom: 0;
